requirements.txt
PyQt6
aiohttp
aiohttp-socks
python-socks
certifi
charset-normalizer
kurigram
tgcrypto
telethon
PySocks
https://github.com/AlreadyNobody/OpenTele-3.13.git@main#egg=OpenTele